pyproject.toml
[project]
name = "automation-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.45",
    "langchain-google-genai>=2.1.8",
    "langgraph>=1.0.9",
    "pypdf2>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "langsmith>=0.7.6",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
